♻️ LocationLog 파일 제거
