+++ app/notice/page.tsx
@@ -6,9 +6,10 @@ export default function NoticePage() {
   const locationLog: Location[] = [PAGES.communication, PAGES.notice];
 
   return (
-    <PageTitle
-      locationLog={locationLog}
-      title="푸른등대 2023gkrsuseh 2학기 학위논문제출기한 단순연장 학위논문제출기한"
-    />
+    <PageTitle locationLog={locationLog}>
+      <h3 className="text-lg font-bold break-keep mr-[55px] font-yoon">
+        푸른등대 2023gkrsuseh 2학기 학위논문제출기한 단순연장 학위논문제출기한
+      </h3>
+    </PageTitle>
   );
 }
